requirements.txt
streamlit
pandas
requests
python-dotenv
yara-python
fuzzywuzzy
python-Levenshtein
networkx
plotly
pyyaml
yara-python
rapidfuzz